ingest/requirements.txt
click==8.1.7
psycopg2-binary==2.9.9
boto3==1.34.7
tenacity==8.2.3